[SYSTEMML-445] [WIP] Prepare GPU backend to support arbitrary datatype
